moving "get_pdos_tj" to the dos module, following pattern used for other spectral functions
