+++ tests/funcs/test_pdos.py
@@ -11,6 +11,31 @@ edata = ElecData(n2_calcdir)
 erange, pdos = get_pdos(edata, idcs=0)
 
 
+def test_get_pdos_tj():
+    from crawfish.core.operations.matrix import get_pdos_tj
+
+    nproj = 3
+    orbs_1 = np.array(list(range(nproj)))
+    orbs_2 = np.array([0,int(nproj-1)])
+    nstates = 3
+    nbands = 3
+
+    proj_tju = (np.random.random([nstates, nbands, nproj]) - 0.5) + 1j*(np.random.random([nstates, nbands, nproj]) - 0.5)
+    wk_t = np.random.random([nstates])
+    pdos_tj_1 = get_pdos_tj(proj_tju, orbs_1, wk_t)
+    pdos_tj_2 = get_pdos_tj(proj_tju, orbs_2, wk_t)
+    for t in range(nstates):
+        for j in range(nbands):
+            assert np.isclose(
+                pdos_tj_1[t,j], 
+                np.sum((np.abs(proj_tju[t,j,orbs_1])**2).flatten())*wk_t[t]
+            )
+            assert np.isclose(
+                pdos_tj_2[t,j], 
+                np.sum((np.abs(proj_tju[t,j,orbs_2])**2).flatten())*wk_t[t]
+            )
+
+
 def test_get_pdos():
     edata = ElecData(EXAMPLE_CALC_DIRS_DIR / "N2_bare_min")
     edata._wk_sabc = np.ones(np.shape(edata.wk_sabc), dtype=REAL_DTYPE)
